fix(validation): updates to schema
1) Fixes problem in Type: validation (where 'Type: enum' was not
called invalid, although 'enum' is not in the list of
simple types) - correctly expressing alternative, mutual exclusive
properties in the schema.

2) Fixes the #/def/url matcher regexp which probably had never been
in correct JS form (required by the JSONSchema spec)

3) Fixes test (001) of the CRDC example MDF, which did not 'assert'
the result of validate_instance_with_schema and so was a no-op.
(once asserted, I found the issue with the URL regexp)

4) Adds test that 'Type: enum' validation is fixed